rename datamodule files
