+++ nugraph/nugraph/data/__init__.py
@@ -1,7 +1,7 @@
 """nugraph.data submodule"""
-from .H5Dataset import NuGraphDataset
-from .H5DataModule import NuGraphDataModule
+from .dataset import NuGraphDataset
+from .data_module import NuGraphDataModule
 
 # legacy imports
-from .H5Dataset import NuGraphDataset as H5Dataset
-from .H5DataModule import NuGraphDataModule as H5DataModule
+from .dataset import NuGraphDataset as H5Dataset
+from .data_module import NuGraphDataModule as H5DataModule
